perf: Use Uint32Array for heatmap generation in AI engine (#9)

+++ src/application/ai/AIEngine.ts
@@ -117,7 +117,7 @@ export class AIEngine {
 
         const width = board.width;
         const height = board.height;
-        const heatMap = new Uint16Array(width * height);
+        const heatMap = new Uint32Array(width * height);
         
         const ITERATIONS = 1000;
